+++ ext/base/dcusumkbn/src/main.c
@@ -88,7 +88,8 @@ void API_SUFFIX(stdlib_strided_dcusumkbn_ndarray)( const CBLAS_INT N, const doub
 			if ( v != 0.0 ) {
 				break;
 			}
-			Y[ iy ] = s + v;
+			s += v;
+			Y[ iy ] = s;
 			ix += strideX;
 			iy += strideY;
 		}