+++ ext/base/gcusumkbn/lib/accessors.js
@@ -93,7 +93,8 @@ function gcusumkbn( N, sum, x, strideX, offsetX, y, strideY, offsetY ) {
 			if ( v !== 0.0 ) {
 				break;
 			}
-			yset( ybuf, iy, s + v );
+			s += v;
+			yset( ybuf, iy, s );
 			ix += strideX;
 			iy += strideY;
 		}