+++ ext/base/gcusumkbn/lib/ndarray.js
@@ -86,7 +86,8 @@ function gcusumkbn( N, sum, x, strideX, offsetX, y, strideY, offsetY ) {
 			if ( v !== 0.0 ) {
 				break;
 			}
-			y[ iy ] = s + v;
+			s += v;
+			y[ iy ] = s;
 			ix += strideX;
 			iy += strideY;
 		}