+++ ext/base/gcusumkbn2/lib/ndarray.js
@@ -87,7 +87,8 @@ function gcusumkbn2( N, sum, x, strideX, offsetX, y, strideY, offsetY ) {
 			if ( v !== 0.0 ) {
 				break;
 			}
-			y[ iy ] = sum + v;
+			sum += v;
+			y[ iy ] = sum;
 			ix += strideX;
 			iy += strideY;
 		}