+++ ext/base/scusumkbn2/src/main.c
@@ -91,7 +91,8 @@ void API_SUFFIX(stdlib_strided_scusumkbn2_ndarray)( const CBLAS_INT N, const flo
 			if ( v != 0.0f ) {
 				break;
 			}
-			Y[ iy ] = s + v;
+			s += v;
+			Y[ iy ] = s;
 			ix += strideX;
 			iy += strideY;
 		}